fix: remvoe wave findDOMNode api
